+++ website/src/AppState.ts
@@ -2,10 +2,8 @@ import { observable } from 'mobx';
 import { InfoRes } from './sdk/server_pb';
 
 class GlobalAppState {
-
   @observable
   info?: InfoRes.AsObject;
-
 }
 
 export const AppState = new GlobalAppState();
@@ -15,5 +13,5 @@ console.info('see global app state by typing "window.AppState"');
 Object.assign(window as any, {
   get AppState() {
     return JSON.parse(JSON.stringify(AppState));
-  }
+  },
 });