+++ website/src/pages/YourDevices.tsx
@@ -3,8 +3,6 @@ import { Devices } from '../components/Devices';
 
 export class YourDevices extends React.Component {
   render() {
-    return (
-      <Devices />
-    );
+    return <Devices />;
   }
 }